fix: keep the rcsbuf lock id a size_t when snapshotting it
rcsbuf_open snapshots the currently-held lock id so it can tell whether the
rcsbuffer already holds one:

    int orig_lockId = rcsbuf->lockId;
    ...
    if(!orig_lockId)
        rcsbuf->lockId = do_lock_file(...);

lockId is a size_t (src/rcs.h:75) and do_lock_file returns size_t
(src/cvs.h:599), so on any 64-bit target the top half is discarded. A lock id
whose low 32 bits are zero - 0x0000000100000000, say - reads as 0, the guard
concludes no lock is held, and a second lock is taken on a file that already
has one. The new id overwrites rcsbuf->lockId, so freercsnode (src/rcs.cpp:766)
releases only the second. The first is held until the process exits, and every
later writer on that ,v blocks on a lock nobody can release.

orig_lockId is read in exactly one place and only as a boolean, so widening
the declaration cannot affect anything else.

Found while reviewing the Tier 1 changes; it sits three lines above them and
is pre-existing, not introduced by that work.

Report: _reports/BUG-server-22-rcsbuf-open-lockid-truncated-to-int.md

+++ cvsnt/cvsnt-2.5.05.3744/src/rcs.cpp
@@ -872,7 +872,7 @@ static void rcsvers_delproc (Node *p)
 static int rcsbuf_open(struct rcsbuffer *rcsbuf, const char *filename)
 {
 	FILE *fp;
-	int orig_lockId = rcsbuf->lockId;
+	size_t orig_lockId = rcsbuf->lockId;
 
 	if(!filename)
 		filename = rcsbuf->filename;
